abotmi/requirements.txt
asgi-redis==1.4.3
boto3==1.4.4
botocore==1.5.70
bs4==0.0.1
celery==4.0.2
channels==1.1.5
Django==1.8.5
django-bootstrap3-datetimepicker==2.2.3
django-debug-toolbar==1.4
django-encrypted-cookie-session==3.0.0
django-longerusernameandemail==0.5.7
django-maintenance-mode==0.4.4
django-mandrill==0.1.0
django-redis==4.8.0
django-simple-history==1.8.1
simplejson==3.8.1
django-storages==1.5.2
django-widget-tweaks==1.4.1
django-wkhtmltopdf==3.1.0
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
html5lib==1.0b8
Fabric==1.13.2
feedparser==5.2.1
MySQL-python==1.2.5
num2words==0.5.4
Pillow==2.9.0
python-wordpress-xmlrpc==2.3
pytz==2015.6
PyYAML==3.11
pyzmq==15.2.0
redis==2.10.5
requests==2.18.1
django-datetime-widget==0.9.3
tld==0.7.6
boto==2.48.0
xhtml2pdf==0.0.6
django-app-namespace-template-loader==0.3
pyethmobisir==0.1
sqlparse==0.1.18 # mandatory for debug toolbar
idna==2.5
urllib3==1.21.1
chardet==3.0.4
certifi==2017.7.27.1
pyOpenSSL==17.3.0
pandas==0.20.1
xlrd==1.1.0
pathlib2==2.1.0
django-mathfilters==0.4.0 #https://pypi.python.org/pypi/django-mathfilters
mongoengine==0.16.3
django-rest-framework-mongoengine==3.3.0